backend/web/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.1
Flask-Script==2.0.5
SQLAlchemy==1.3.2
psycopg2==2.8.1
Passlib==1.7.1
pyjwt==1.4.2
coverage==4.2
Flask-Migrate==2.0.2
